refactor(subway): rename classes
- 클래스명 변경
