+++ src/main/java/subway/service/line/LineAddingService.java
@@ -6,27 +6,10 @@ import subway.domain.TransitMap;
 import subway.repository.LineRepository;
 import subway.repository.TransitMapRepository;
 import subway.view.output.line.LineInformationView;
-import subway.view.output.line.LineTextView;
 
 import java.util.LinkedList;
-import java.util.Scanner;
 
 public class LineAddingService {
-    public static String scanLineName(Scanner scanner) {
-        LineTextView.printLineAddingText();
-        return scanner.nextLine();
-    }
-
-    public static String scanUpStationName(Scanner scanner) {
-        LineTextView.printLineUpStationNameText();
-        return scanner.nextLine();
-    }
-
-    public static String scanDownStationName(Scanner scanner) {
-        LineTextView.printLineDownStationNameText();
-        return scanner.nextLine();
-    }
-
     public static LinkedList<Station> addStationNames(String upStationName, String downStationName) {
         LinkedList<Station> stationNames = new LinkedList<>();
