+++ src/main/java/subway/service/section/SectionScannerService.java
@@ -1,0 +1,22 @@
+package subway.service.section;
+
+import subway.view.output.section.SectionTextView;
+
+import java.util.Scanner;
+
+public class SectionScannerService {
+    public static String scanLineName(Scanner scanner) {
+        SectionTextView.printSectionAddingLineText();
+        return scanner.nextLine();
+    }
+
+    public static String scanStationName(Scanner scanner) {
+        SectionTextView.printSectionAddingStationText();
+        return scanner.nextLine();
+    }
+
+    public static String scanOrder(Scanner scanner) {
+        SectionTextView.printSectionAddingOrderText();
+        return scanner.nextLine();
+    }
+}
